Work in progress
Neutron can now decay into Protons and Electrons. Need to figure out a good way to test this functionality. Currently the test fails because of the previously run tests, because the universe is a Singleton. Even after resetting the universe I still have the same issue, and that is because of the race condition caused by the neutrons that were created in previous tests. Also need to think about overriding the time that it takes to decay for testing. But I would hate to add that code to the production data model, so I am not completely sure what to do for this.

Also moved neutron into the subatomic particle project, there were too many projects and it was quickly growing. Thinking that all of the subatomic particles will sit in this project. Then the universe will have its own project. That might be all of the projects besides any applications or tests that reference these libraries.