Figured out how to properly do the test for neutron decay.
I made the timer on the neutron public so that I can modify the timer during the test. This enables the production code to continue to have the standard timer and no code that updates the timer to a different time in production code either. I then created a utility that can be given a timer and it will cause that timer to trigger in one millisecond, and then pause the application for five hundred milliseconds.