Working on perfecting the behavior
Updated CompositeParticle to add the new particle to the universe upon its creation. I realize at this point that I will need to move that to SubatomicParticle, but I do not yet have an implementation for that. Added a test that will make sure that neutron is added to the universe at creation. Also made an abstract class that I will hook up to the other subatomic particle tests. Finally, I noticed that I made the Electron a CompositeParticle when in reality it should have been an ElementalParticle, so I will have to update that in a future update.