Introduced additional abstract classes
Now electrons know that they are actually elementary particles versus composite particles. This was done by adding an abstract class ElementaryParticle. This class needed to share some of the logic that existed in CompositeParticle so I made a common abstract class that they both extend, SubatomicParticle. This is similar to how the interfaces are setup. Finally it appears that I forgot to check-in the changes to the ElectronTests and ProtonTests from the previous commit, so they are included in here as well.