Introducing Quarks.
Had to introduce tracking mass in electron volts as I have not been able to find articles that state the mass of quarks in kilograms. I did manage to find some converters that would convert between electron volts and kilograms, but they either were not precise enough or contained rounding errors. I was able to prove this as the Wikipedia pages for electrons, protons, and neutrons provided the mass for those particles in both measurements. The converters were unable to convert either number and produce the expected result. Therefore I just added the mass in electron volts and set the original mass to be mass in kilograms now. Both masses are also nullable now just in case I don't happen to have the required data, as in the case of the quarks. Currently mass does not do anything, but at some point it will once I introduce gravitational force.

There are six types of quarks. This first bit of code is adding Up and Down quarks, these are required to make neutrons and protons. Also managed to create a generic unit test, which pleases me greatly, see QuarkTests.