Introducing the concept of Hadrons and Baryons. Also Neutrons contain quarks now.
Moved some folders around, did this in order to have a visual representation of all the children and their parents. The relationships are getting quite deep already and I suspect it will only get deeper as new subatomic particles are added. For example a Neutron is now a Baryon, which is a Hadron, which is a Composite Particle, which is a Subatomic Particle.

Since the Neutron is now made up of Quarks I am now testing to make that during creation it contains the Quarks that we expect. I also realized that I was not testing for its charge value or mass in electron volts. I think this occurred because the Neutron predated those concepts. But adding the test showed a glaring error where I had set the constant for a Neutron's charge to be a one (1) rather than a zero (0). How embarrassing.

Baryons require at least three quarks, so I will throw an error if one is attempted to be made without them. Once again Neutrons are now Baryons, Protons will follow soon but this commit does not contain that change.