Fixed Neutron decay test
Quarks are now subatomic particles in the universe, and that changes the count of the particles for this test. Also it was a rather bad idea to check that the universe only contained a set number of particles rather than checking to see if the particles were in the universe.

Found that with LINQ you can filter a collection by using OfType, and it will return all of the elements of that type.