Added the strong nuclear force
Gluons are bound to Quarks through the strong nuclear force. This is accomplished by creating a timer on the creation of a Gluon. This timer will go off each second looking for a Quark to bind to until it does so. A single Gluon can bind two Quarks together so I created two timer events, one for each Quark.

Subatomic particles now know if they have been bound to another object. Later on this will be helpful for capturing Neutrons, Protons, and Electrons. Neutrons and Protons are bound through the nuclear force, Electrons will be bound to Protons through the electrical force. I think the code will be similar at this point though, since distance and mass are not being considered.