Added Epochs, Time, and Temperature to the Universe
This is super important to simulate how certain things will behave differently at various stages of the Universe. For example, and it is coded in this commit, the Universe is too hot until the tail end of the Quark Epoch to have Quarks and Gluons bind together through the Strong Nuclear Force.

Once the universe cools to one trillion kelvin the Strong Nuclear Force is able to begin converting the Quark Gluon Plasma into Quarks that are bound to Gluons.