Updated the story of the Universe
Story now begins at the Planck epoch and moves into the Quark epoch to form the first Quarks and Gluons. We then end at the transition from the end of the Quark epoch to the Hadron epoch, still need to code for the Hadron epoch, where Quarks and Gluons combine to form a Hadron. Then I make a Neutron out of the bound Quarks and Gluons. From there the story progresses as it did before.

I have noticed that there is a bug where the Quarks and Gluons are added back to the Universe after I get my Neutron returned. But I am not sure if this occurs when I attach the Gluons and the Quarks or when the Neutron is created. Will need to do some testing and debugging to track this down.