Found the bug and it is squashed.
Turns out that when I initialized the Neutron class for the first time it also initializes the static fields on that class. That caused the Quarks from ConstantComposition to be added to the universe. Strangely the same did not happen with the Gluons from ConstantGluons, I think this is because I was not referencing the Baryon's class directly. To fix this I turned ConstantComposition into a function.

I updated the Neutron tests to use the new code and also cleaned it up some, used a variable to store the quarks versus retrieving them one by one. This also caused the tests to fail and required me to introduce some equality logic into the solution.