WIP. Currently adding in Antiparticles and Mesons.
The Meson is comprised of Quarks and Antiquarks, when these two collide then they will annihilate one another and will decay into other subatomic particles. These particles will change based on the type of Meson, first one I will be coding for will be a Pion, which will decay into a Muon, which decays into an Electron. Which is the entire reason I am going down this road, so that I can correctly generate Electrons for my Hydrogen atom.

For the antiparticles I am currently thinking that I will just store the type of the particle on its opposite, this way it will know what to look out for but it will not necessarily be a specific object. This mimics real life where any particle-antiparticle will annihilate versus just specific instances of each.