Finished adding antiparticles
Wow this was a whole bunch of work just so that I can begin working on creating Pions (Pi Mesons). This particle requires an UpQuark and a AntiDownQuark. Pions are required as they decay into Muons, and Muons will decay into Electrons. I also added some logic to the CompositeParticle data model to make sure that a particle has a charge of -1, 0, or 1. I think I am correct in saying that all composite particles have a charge that equals one of those three numbers. Also added the ability to create a Proton with specific Quarks and Gluons. This feature will be important at a later date when I update the beta minus decay of the Neutron. I plan on taking one of the DownQuarks and converting it to an UpQuark and then using the Quarks and Gluon from the Neutron in the creation of the Proton. This will more accurately simulate what is occuring in the beta minus decay.