Discovered that there are three versions of Pions. I have created those now.
Mesons now has its own Gluon constant, this matches what Baryons do for composite particles. Then I created the abstract class Pion and the classes for the three versions of Pion. There will most likely some additional code that I will need to write to handle the decay mechanism, according to Wikipedia there are a lot of options for the outcome of the decay, and it might be a probabilistic thing, should be fun to code.