Testing new Quark data models.
First thing that I did was to make a common validation class for the creation of the Quarks. It was one thing to have the code in two places, but in total there are twelve Quark flavors to support and that would be a gross copy-paste exercise. This new way is much cleaner. From there it was just a simple task of updating the Up and Down Quarks to use the new method, and then create the test classes for AntiUp and AntiDown. Also realized that AntiDown data model was originally coded incorrectly as I had set it to have a QuarkFlavor of Up. Only noticed this while reading the test where I had set it to Up as well, consistency in that mistake, and then I realized that the test should have failed. Finally updated Quark equality to test for the Antiparticle as well.