Work in progress.
I realized that I reintroduced a bug that I had already solved and I am beginning to write some tests that will prevent this from happening again. Hard thing with this is that the bug is very hard to test due to the Universe being a Singleton. My solution for this is to introduce dependency injection and create an isolated universe for tests that require it.