Work in progress.
Didn't have much time for this session. But I am thinking that I will make an instance method on the interface that will return the expected Universe. Feels gross, but it might be "the best of our bad plans". Currently writing what feels like an essay in the comments, and so I have put a bunch of notes to remember what my current thoughts are on this, will fill it all in later on.