Work in progress.
OK. So I think I finally have a viable solution, that I actually like, for the Universe Singleton issue. You can see previous commits where I tried to add dependency injection to allow me to use different data models for the Universe. That works fine until it comes to particle creation, where the particles were coded in such a way that a Singleton is expected to exist. So that was a bummer. I briefly considered rewriting a huge chunk of this code to require that you first create a Universe and then manually add particles to it, where all of the creation logic would fall into the Universe. I really just hate that idea, and it seems so poorly designed that I just did not want to code for it. But then I thought about using some type of event system where the Universe can subscribe to particle creation and then automatically add the particles to the Universe when they spawn. This would allow me to decouple the SubatomicParticle class from the Universe entirely.

So I started writing this code but then ran into another dilemma, I could not call a virtual method from within a constructor. Of course I need the event trigger to be a virtual method, so that I can override its logic in derived classes, so this left me stuck for a short while. Then I stumbled upon a StackOverflow post that mentioned using the Factory Design pattern. That is where I currently am. I have implemented a portion of the pattern on the SubatomicParticle and UpQuark classes, and I have written a unit test that I think will test it. Need to implement this on all of the other data models in order to begin running tests and see how this works out. Then I will write unit tests to check that the event is fired. At some point this will be clean again, but currently will be in a bit of a disarray as I take it all apart and put it back together again.