Massive refactoring is complete. The Universe now adds a particle to itself upon particle creation.
Moved all of the Subatomic Particle project files to the Universe project, this was so that the Universe could reference the SubatomicParticleCreator when registering to listen to the event. I could have referenced the SubatomicParticle project but that would have created a circular reference due to Neutron, will need to clean up that file at a later time.